Upgrade to 2013-08-15, CreateTable, Cors Rules
Also split queue tests into separate files with a utils object, will use this pattern for other tests now too
