+++ Basic.Azure.Storage/Basic.Azure.Storage.csproj
@@ -33,6 +33,10 @@
     <Reference Include="Microsoft.Practices.TransientFaultHandling.Core">
       <HintPath>..\packages\TransientFaultHandling.Core.5.1.1209.1\lib\NET4\Microsoft.Practices.TransientFaultHandling.Core.dll</HintPath>
     </Reference>
+    <Reference Include="Newtonsoft.Json, Version=9.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
+      <HintPath>..\packages\Newtonsoft.Json.9.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -121,6 +125,7 @@
     <Compile Include="Communications\Common\BlobSignedIdentifier.cs" />
     <Compile Include="Communications\Common\QueueSignedIdentifier.cs" />
     <Compile Include="Communications\Common\StorageAnalyticsVersionNumber.cs" />
+    <Compile Include="Communications\Common\StorageServiceCorsRule.cs" />
     <Compile Include="Communications\Common\StorageServiceLoggingProperties.cs" />
     <Compile Include="Communications\Common\StorageServiceMetricsProperties.cs" />
     <Compile Include="Communications\Common\StorageServiceProperties.cs" />
@@ -188,8 +193,10 @@
     </Compile>
     <Compile Include="Communications\ServiceExceptions\UnidentifiedAzureException.cs" />
     <Compile Include="Communications\ServiceExceptions\UnrecognizedAzureException.cs" />
-    <Compile Include="Communications\TableService\CreateTableRequest.cs" />
-    <Compile Include="Communications\TableService\CreateTableResponse.cs" />
+    <Compile Include="Communications\TableService\EntityOperations\InsertEntityRequest.cs" />
+    <Compile Include="Communications\TableService\TableOperations\CreateTableRequest.cs" />
+    <Compile Include="Communications\TableService\TableOperations\CreateTableResponse.cs" />
+    <Compile Include="Communications\TableService\Interfaces\ITableEntity.cs" />
     <Compile Include="Communications\TableService\MetadataPreference.cs" />
     <Compile Include="Communications\Utility\Base64Converter.cs" />
     <Compile Include="Communications\Utility\BlobListIncludeParse.cs" />
