+++ Basic.Azure.Storage/Communications/TableService/TableOperations/CreateTableResponse.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Linq;
 
-namespace Basic.Azure.Storage.Communications.TableService
+namespace Basic.Azure.Storage.Communications.TableService.TableOperations
 {
     public class CreateTableResponse : IResponsePayload, IReceiveAdditionalHeadersWithResponse, IReceiveDataWithResponse
     {
